feat : metadata 수정 , favicon 수정
